+++ planner/core/common_plans.go
@@ -50,6 +50,7 @@ import (
 	"github.com/pingcap/tidb/util/hint"
 	"github.com/pingcap/tidb/util/kvcache"
 	"github.com/pingcap/tidb/util/logutil"
+	"github.com/pingcap/tidb/util/ranger"
 	"github.com/pingcap/tidb/util/texttree"
 	"github.com/tikv/client-go/v2/oracle"
 	"go.uber.org/zap"
@@ -405,6 +406,12 @@ func (e *Execute) checkPreparedPriv(ctx context.Context, sctx sessionctx.Context
 	return err
 }
 
+func (e *Execute) setFoundInPlanCache(sctx sessionctx.Context, opt bool) error {
+	vars := sctx.GetSessionVars()
+	err := vars.SetSystemVar(variable.TiDBFoundInPlanCache, variable.BoolToOnOff(opt))
+	return err
+}
+
 // GetBindSQL4PlanCache used to get the bindSQL for plan cache to build the plan cache key.
 func GetBindSQL4PlanCache(sctx sessionctx.Context, preparedStmt *CachedPrepareStmt) string {
 	useBinding := sctx.GetSessionVars().UsePlanBaselines
@@ -491,7 +498,7 @@ func (e *Execute) getPhysicalPlan(ctx context.Context, sctx sessionctx.Context, 
 		// so you don't need to consider whether prepared.useCache is enabled.
 		plan := prepared.CachedPlan.(Plan)
 		names := prepared.CachedNames.(types.NameSlice)
-		err := RebuildPlan(plan)
+		err := e.RebuildPlan(plan)
 		if err != nil {
 			logutil.BgLogger().Debug("rebuild range failed", zap.Error(err))
 			goto REBUILD
@@ -501,7 +508,7 @@ func (e *Execute) getPhysicalPlan(ctx context.Context, sctx sessionctx.Context, 
 		} else {
 			planCacheCounter.Inc()
 		}
-		err = setFoundInPlanCache(sctx, true)
+		err = e.setFoundInPlanCache(sctx, true)
 		if err != nil {
 			return err
 		}
@@ -538,12 +545,12 @@ func (e *Execute) getPhysicalPlan(ctx context.Context, sctx sessionctx.Context, 
 					}
 				}
 				if planValid {
-					err := RebuildPlan(cachedVal.Plan)
+					err := e.RebuildPlan(cachedVal.Plan)
 					if err != nil {
 						logutil.BgLogger().Debug("rebuild range failed", zap.Error(err))
 						goto REBUILD
 					}
-					err = setFoundInPlanCache(sctx, true)
+					err = e.setFoundInPlanCache(sctx, true)
 					if err != nil {
 						return err
 					}
@@ -617,7 +624,7 @@ REBUILD:
 			sctx.PreparedPlanCache().Put(cacheKey, []*PlanCacheValue{cached})
 		}
 	}
-	err = setFoundInPlanCache(sctx, false)
+	err = e.setFoundInPlanCache(sctx, false)
 	return err
 }
 
@@ -668,6 +675,214 @@ func (e *Execute) tryCachePointPlan(ctx context.Context, sctx sessionctx.Context
 	return err
 }
 
+// RebuildPlan will rebuild this plan under current user parameters.
+func (e *Execute) RebuildPlan(p Plan) error {
+	sc := p.SCtx().GetSessionVars().StmtCtx
+	sc.InPreparedPlanBuilding = true
+	defer func() { sc.InPreparedPlanBuilding = false }()
+	return e.rebuildRange(p)
+}
+
+func (e *Execute) rebuildRange(p Plan) error {
+	sctx := p.SCtx()
+	sc := p.SCtx().GetSessionVars().StmtCtx
+	var err error
+	switch x := p.(type) {
+	case *PhysicalIndexHashJoin:
+		return e.rebuildRange(&x.PhysicalIndexJoin)
+	case *PhysicalIndexMergeJoin:
+		return e.rebuildRange(&x.PhysicalIndexJoin)
+	case *PhysicalIndexJoin:
+		if err := x.Ranges.Rebuild(); err != nil {
+			return err
+		}
+		for _, child := range x.Children() {
+			err = e.rebuildRange(child)
+			if err != nil {
+				return err
+			}
+		}
+	case *PhysicalTableScan:
+		err = e.buildRangeForTableScan(sctx, x)
+		if err != nil {
+			return err
+		}
+	case *PhysicalIndexScan:
+		err = e.buildRangeForIndexScan(sctx, x)
+		if err != nil {
+			return err
+		}
+	case *PhysicalTableReader:
+		err = e.rebuildRange(x.TablePlans[0])
+		if err != nil {
+			return err
+		}
+	case *PhysicalIndexReader:
+		err = e.rebuildRange(x.IndexPlans[0])
+		if err != nil {
+			return err
+		}
+	case *PhysicalIndexLookUpReader:
+		err = e.rebuildRange(x.IndexPlans[0])
+		if err != nil {
+			return err
+		}
+	case *PointGetPlan:
+		// if access condition is not nil, which means it's a point get generated by cbo.
+		if x.AccessConditions != nil {
+			if x.IndexInfo != nil {
+				ranges, err := ranger.DetachCondAndBuildRangeForIndex(x.ctx, x.AccessConditions, x.IdxCols, x.IdxColLens)
+				if err != nil {
+					return err
+				}
+				if len(ranges.Ranges) == 0 || len(ranges.AccessConds) != len(x.AccessConditions) {
+					return errors.New("failed to rebuild range: the length of the range has changed")
+				}
+				for i := range x.IndexValues {
+					x.IndexValues[i] = ranges.Ranges[0].LowVal[i]
+				}
+			} else {
+				var pkCol *expression.Column
+				if x.TblInfo.PKIsHandle {
+					if pkColInfo := x.TblInfo.GetPkColInfo(); pkColInfo != nil {
+						pkCol = expression.ColInfo2Col(x.schema.Columns, pkColInfo)
+					}
+				}
+				if pkCol != nil {
+					ranges, err := ranger.BuildTableRange(x.AccessConditions, x.ctx, pkCol.RetType)
+					if err != nil {
+						return err
+					}
+					if len(ranges) == 0 {
+						return errors.New("failed to rebuild range: the length of the range has changed")
+					}
+					x.Handle = kv.IntHandle(ranges[0].LowVal[0].GetInt64())
+				}
+			}
+		}
+		// The code should never run here as long as we're not using point get for partition table.
+		// And if we change the logic one day, here work as defensive programming to cache the error.
+		if x.PartitionInfo != nil {
+			// TODO: relocate the partition after rebuilding range to make PlanCache support PointGet
+			return errors.New("point get for partition table can not use plan cache")
+		}
+		if x.HandleConstant != nil {
+			dVal, err := convertConstant2Datum(sc, x.HandleConstant, x.handleFieldType)
+			if err != nil {
+				return err
+			}
+			iv, err := dVal.ToInt64(sc)
+			if err != nil {
+				return err
+			}
+			x.Handle = kv.IntHandle(iv)
+			return nil
+		}
+		for i, param := range x.IndexConstants {
+			if param != nil {
+				dVal, err := convertConstant2Datum(sc, param, x.ColsFieldType[i])
+				if err != nil {
+					return err
+				}
+				x.IndexValues[i] = *dVal
+			}
+		}
+		return nil
+	case *BatchPointGetPlan:
+		// if access condition is not nil, which means it's a point get generated by cbo.
+		if x.AccessConditions != nil {
+			if x.IndexInfo != nil {
+				ranges, err := ranger.DetachCondAndBuildRangeForIndex(x.ctx, x.AccessConditions, x.IdxCols, x.IdxColLens)
+				if err != nil {
+					return err
+				}
+				if len(ranges.Ranges) != len(x.IndexValues) || len(ranges.AccessConds) != len(x.AccessConditions) {
+					return errors.New("failed to rebuild range: the length of the range has changed")
+				}
+				for i := range x.IndexValues {
+					copy(x.IndexValues[i], ranges.Ranges[i].LowVal)
+				}
+			} else {
+				var pkCol *expression.Column
+				if x.TblInfo.PKIsHandle {
+					if pkColInfo := x.TblInfo.GetPkColInfo(); pkColInfo != nil {
+						pkCol = expression.ColInfo2Col(x.schema.Columns, pkColInfo)
+					}
+				}
+				if pkCol != nil {
+					ranges, err := ranger.BuildTableRange(x.AccessConditions, x.ctx, pkCol.RetType)
+					if err != nil {
+						return err
+					}
+					if len(ranges) != len(x.Handles) {
+						return errors.New("failed to rebuild range: the length of the range has changed")
+					}
+					for i := range ranges {
+						x.Handles[i] = kv.IntHandle(ranges[i].LowVal[0].GetInt64())
+					}
+				}
+			}
+		}
+		for i, param := range x.HandleParams {
+			if param != nil {
+				dVal, err := convertConstant2Datum(sc, param, x.HandleType)
+				if err != nil {
+					return err
+				}
+				iv, err := dVal.ToInt64(sc)
+				if err != nil {
+					return err
+				}
+				x.Handles[i] = kv.IntHandle(iv)
+			}
+		}
+		for i, params := range x.IndexValueParams {
+			if len(params) < 1 {
+				continue
+			}
+			for j, param := range params {
+				if param != nil {
+					dVal, err := convertConstant2Datum(sc, param, x.IndexColTypes[j])
+					if err != nil {
+						return err
+					}
+					x.IndexValues[i][j] = *dVal
+				}
+			}
+		}
+	case *PhysicalIndexMergeReader:
+		indexMerge := p.(*PhysicalIndexMergeReader)
+		for _, partialPlans := range indexMerge.PartialPlans {
+			err = e.rebuildRange(partialPlans[0])
+			if err != nil {
+				return err
+			}
+		}
+		// We don't need to handle the indexMerge.TablePlans, because the tablePlans
+		// only can be (Selection) + TableRowIDScan. There have no range need to rebuild.
+	case PhysicalPlan:
+		for _, child := range x.Children() {
+			err = e.rebuildRange(child)
+			if err != nil {
+				return err
+			}
+		}
+	case *Insert:
+		if x.SelectPlan != nil {
+			return e.rebuildRange(x.SelectPlan)
+		}
+	case *Update:
+		if x.SelectPlan != nil {
+			return e.rebuildRange(x.SelectPlan)
+		}
+	case *Delete:
+		if x.SelectPlan != nil {
+			return e.rebuildRange(x.SelectPlan)
+		}
+	}
+	return nil
+}
+
 func convertConstant2Datum(sc *stmtctx.StatementContext, con *expression.Constant, target *types.FieldType) (*types.Datum, error) {
 	val, err := con.Eval(chunk.Row{})
 	if err != nil {
@@ -683,6 +898,73 @@ func convertConstant2Datum(sc *stmtctx.StatementContext, con *expression.Constan
 		return nil, errors.New("Convert constant to datum is failed, because the constant has changed after the covert")
 	}
 	return &dVal, nil
+}
+
+func (e *Execute) buildRangeForTableScan(sctx sessionctx.Context, ts *PhysicalTableScan) (err error) {
+	if ts.Table.IsCommonHandle {
+		pk := tables.FindPrimaryIndex(ts.Table)
+		pkCols := make([]*expression.Column, 0, len(pk.Columns))
+		pkColsLen := make([]int, 0, len(pk.Columns))
+		for _, colInfo := range pk.Columns {
+			if pkCol := expression.ColInfo2Col(ts.schema.Columns, ts.Table.Columns[colInfo.Offset]); pkCol != nil {
+				pkCols = append(pkCols, pkCol)
+				// We need to consider the prefix index.
+				// For example: when we have 'a varchar(50), index idx(a(10))'
+				// So we will get 'colInfo.Length = 50' and 'pkCol.RetType.flen = 10'.
+				// In 'hasPrefix' function from 'util/ranger/ranger.go' file,
+				// we use 'columnLength == types.UnspecifiedLength' to check whether we have prefix index.
+				if colInfo.Length != types.UnspecifiedLength && colInfo.Length == pkCol.RetType.GetFlen() {
+					pkColsLen = append(pkColsLen, types.UnspecifiedLength)
+				} else {
+					pkColsLen = append(pkColsLen, colInfo.Length)
+				}
+			}
+		}
+		if len(pkCols) > 0 {
+			res, err := ranger.DetachCondAndBuildRangeForIndex(sctx, ts.AccessCondition, pkCols, pkColsLen)
+			if err != nil {
+				return err
+			}
+			if len(res.AccessConds) != len(ts.AccessCondition) {
+				return errors.New("rebuild range for cached plan failed")
+			}
+			ts.Ranges = res.Ranges
+		} else {
+			ts.Ranges = ranger.FullRange()
+		}
+	} else {
+		var pkCol *expression.Column
+		if ts.Table.PKIsHandle {
+			if pkColInfo := ts.Table.GetPkColInfo(); pkColInfo != nil {
+				pkCol = expression.ColInfo2Col(ts.schema.Columns, pkColInfo)
+			}
+		}
+		if pkCol != nil {
+			ts.Ranges, err = ranger.BuildTableRange(ts.AccessCondition, sctx, pkCol.RetType)
+			if err != nil {
+				return err
+			}
+		} else {
+			ts.Ranges = ranger.FullIntRange(false)
+		}
+	}
+	return
+}
+
+func (e *Execute) buildRangeForIndexScan(sctx sessionctx.Context, is *PhysicalIndexScan) (err error) {
+	if len(is.IdxCols) == 0 {
+		is.Ranges = ranger.FullRange()
+		return
+	}
+	res, err := ranger.DetachCondAndBuildRangeForIndex(sctx, is.AccessCondition, is.IdxCols, is.IdxColLens)
+	if err != nil {
+		return err
+	}
+	if len(res.AccessConds) != len(is.AccessCondition) {
+		return errors.New("rebuild range for cached plan failed")
+	}
+	is.Ranges = res.Ranges
+	return
 }
 
 // Deallocate represents deallocate plan.
