+++ planner/core/plan_cache.go
@@ -17,15 +17,9 @@ package core
 import (
 	"context"
 
-	"github.com/pingcap/errors"
-	"github.com/pingcap/tidb/expression"
-	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/parser/ast"
 	"github.com/pingcap/tidb/sessionctx"
-	"github.com/pingcap/tidb/sessionctx/variable"
-	"github.com/pingcap/tidb/table/tables"
 	"github.com/pingcap/tidb/types"
-	"github.com/pingcap/tidb/util/ranger"
 	"github.com/pingcap/tidb/util/sqlexec"
 )
 
@@ -70,285 +64,4 @@ func getPlanFromCache(ctx context.Context, sctx sessionctx.Context, handler prep
 	}
 
 	return rs, true, nil
-}
-
-// RebuildPlan will rebuild this plan under current user parameters.
-func RebuildPlan(p Plan) error {
-	sc := p.SCtx().GetSessionVars().StmtCtx
-	sc.InPreparedPlanBuilding = true
-	defer func() { sc.InPreparedPlanBuilding = false }()
-	return rebuildRange(p)
-}
-
-func rebuildRange(p Plan) error {
-	sctx := p.SCtx()
-	sc := p.SCtx().GetSessionVars().StmtCtx
-	var err error
-	switch x := p.(type) {
-	case *PhysicalIndexHashJoin:
-		return rebuildRange(&x.PhysicalIndexJoin)
-	case *PhysicalIndexMergeJoin:
-		return rebuildRange(&x.PhysicalIndexJoin)
-	case *PhysicalIndexJoin:
-		if err := x.Ranges.Rebuild(); err != nil {
-			return err
-		}
-		for _, child := range x.Children() {
-			err = rebuildRange(child)
-			if err != nil {
-				return err
-			}
-		}
-	case *PhysicalTableScan:
-		err = buildRangeForTableScan(sctx, x)
-		if err != nil {
-			return err
-		}
-	case *PhysicalIndexScan:
-		err = buildRangeForIndexScan(sctx, x)
-		if err != nil {
-			return err
-		}
-	case *PhysicalTableReader:
-		err = rebuildRange(x.TablePlans[0])
-		if err != nil {
-			return err
-		}
-	case *PhysicalIndexReader:
-		err = rebuildRange(x.IndexPlans[0])
-		if err != nil {
-			return err
-		}
-	case *PhysicalIndexLookUpReader:
-		err = rebuildRange(x.IndexPlans[0])
-		if err != nil {
-			return err
-		}
-	case *PointGetPlan:
-		// if access condition is not nil, which means it's a point get generated by cbo.
-		if x.AccessConditions != nil {
-			if x.IndexInfo != nil {
-				ranges, err := ranger.DetachCondAndBuildRangeForIndex(x.ctx, x.AccessConditions, x.IdxCols, x.IdxColLens)
-				if err != nil {
-					return err
-				}
-				if len(ranges.Ranges) == 0 || len(ranges.AccessConds) != len(x.AccessConditions) {
-					return errors.New("failed to rebuild range: the length of the range has changed")
-				}
-				for i := range x.IndexValues {
-					x.IndexValues[i] = ranges.Ranges[0].LowVal[i]
-				}
-			} else {
-				var pkCol *expression.Column
-				if x.TblInfo.PKIsHandle {
-					if pkColInfo := x.TblInfo.GetPkColInfo(); pkColInfo != nil {
-						pkCol = expression.ColInfo2Col(x.schema.Columns, pkColInfo)
-					}
-				}
-				if pkCol != nil {
-					ranges, err := ranger.BuildTableRange(x.AccessConditions, x.ctx, pkCol.RetType)
-					if err != nil {
-						return err
-					}
-					if len(ranges) == 0 {
-						return errors.New("failed to rebuild range: the length of the range has changed")
-					}
-					x.Handle = kv.IntHandle(ranges[0].LowVal[0].GetInt64())
-				}
-			}
-		}
-		// The code should never run here as long as we're not using point get for partition table.
-		// And if we change the logic one day, here work as defensive programming to cache the error.
-		if x.PartitionInfo != nil {
-			// TODO: relocate the partition after rebuilding range to make PlanCache support PointGet
-			return errors.New("point get for partition table can not use plan cache")
-		}
-		if x.HandleConstant != nil {
-			dVal, err := convertConstant2Datum(sc, x.HandleConstant, x.handleFieldType)
-			if err != nil {
-				return err
-			}
-			iv, err := dVal.ToInt64(sc)
-			if err != nil {
-				return err
-			}
-			x.Handle = kv.IntHandle(iv)
-			return nil
-		}
-		for i, param := range x.IndexConstants {
-			if param != nil {
-				dVal, err := convertConstant2Datum(sc, param, x.ColsFieldType[i])
-				if err != nil {
-					return err
-				}
-				x.IndexValues[i] = *dVal
-			}
-		}
-		return nil
-	case *BatchPointGetPlan:
-		// if access condition is not nil, which means it's a point get generated by cbo.
-		if x.AccessConditions != nil {
-			if x.IndexInfo != nil {
-				ranges, err := ranger.DetachCondAndBuildRangeForIndex(x.ctx, x.AccessConditions, x.IdxCols, x.IdxColLens)
-				if err != nil {
-					return err
-				}
-				if len(ranges.Ranges) != len(x.IndexValues) || len(ranges.AccessConds) != len(x.AccessConditions) {
-					return errors.New("failed to rebuild range: the length of the range has changed")
-				}
-				for i := range x.IndexValues {
-					copy(x.IndexValues[i], ranges.Ranges[i].LowVal)
-				}
-			} else {
-				var pkCol *expression.Column
-				if x.TblInfo.PKIsHandle {
-					if pkColInfo := x.TblInfo.GetPkColInfo(); pkColInfo != nil {
-						pkCol = expression.ColInfo2Col(x.schema.Columns, pkColInfo)
-					}
-				}
-				if pkCol != nil {
-					ranges, err := ranger.BuildTableRange(x.AccessConditions, x.ctx, pkCol.RetType)
-					if err != nil {
-						return err
-					}
-					if len(ranges) != len(x.Handles) {
-						return errors.New("failed to rebuild range: the length of the range has changed")
-					}
-					for i := range ranges {
-						x.Handles[i] = kv.IntHandle(ranges[i].LowVal[0].GetInt64())
-					}
-				}
-			}
-		}
-		for i, param := range x.HandleParams {
-			if param != nil {
-				dVal, err := convertConstant2Datum(sc, param, x.HandleType)
-				if err != nil {
-					return err
-				}
-				iv, err := dVal.ToInt64(sc)
-				if err != nil {
-					return err
-				}
-				x.Handles[i] = kv.IntHandle(iv)
-			}
-		}
-		for i, params := range x.IndexValueParams {
-			if len(params) < 1 {
-				continue
-			}
-			for j, param := range params {
-				if param != nil {
-					dVal, err := convertConstant2Datum(sc, param, x.IndexColTypes[j])
-					if err != nil {
-						return err
-					}
-					x.IndexValues[i][j] = *dVal
-				}
-			}
-		}
-	case *PhysicalIndexMergeReader:
-		indexMerge := p.(*PhysicalIndexMergeReader)
-		for _, partialPlans := range indexMerge.PartialPlans {
-			err = rebuildRange(partialPlans[0])
-			if err != nil {
-				return err
-			}
-		}
-		// We don't need to handle the indexMerge.TablePlans, because the tablePlans
-		// only can be (Selection) + TableRowIDScan. There have no range need to rebuild.
-	case PhysicalPlan:
-		for _, child := range x.Children() {
-			err = rebuildRange(child)
-			if err != nil {
-				return err
-			}
-		}
-	case *Insert:
-		if x.SelectPlan != nil {
-			return rebuildRange(x.SelectPlan)
-		}
-	case *Update:
-		if x.SelectPlan != nil {
-			return rebuildRange(x.SelectPlan)
-		}
-	case *Delete:
-		if x.SelectPlan != nil {
-			return rebuildRange(x.SelectPlan)
-		}
-	}
-	return nil
-}
-
-func buildRangeForTableScan(sctx sessionctx.Context, ts *PhysicalTableScan) (err error) {
-	if ts.Table.IsCommonHandle {
-		pk := tables.FindPrimaryIndex(ts.Table)
-		pkCols := make([]*expression.Column, 0, len(pk.Columns))
-		pkColsLen := make([]int, 0, len(pk.Columns))
-		for _, colInfo := range pk.Columns {
-			if pkCol := expression.ColInfo2Col(ts.schema.Columns, ts.Table.Columns[colInfo.Offset]); pkCol != nil {
-				pkCols = append(pkCols, pkCol)
-				// We need to consider the prefix index.
-				// For example: when we have 'a varchar(50), index idx(a(10))'
-				// So we will get 'colInfo.Length = 50' and 'pkCol.RetType.flen = 10'.
-				// In 'hasPrefix' function from 'util/ranger/ranger.go' file,
-				// we use 'columnLength == types.UnspecifiedLength' to check whether we have prefix index.
-				if colInfo.Length != types.UnspecifiedLength && colInfo.Length == pkCol.RetType.GetFlen() {
-					pkColsLen = append(pkColsLen, types.UnspecifiedLength)
-				} else {
-					pkColsLen = append(pkColsLen, colInfo.Length)
-				}
-			}
-		}
-		if len(pkCols) > 0 {
-			res, err := ranger.DetachCondAndBuildRangeForIndex(sctx, ts.AccessCondition, pkCols, pkColsLen)
-			if err != nil {
-				return err
-			}
-			if len(res.AccessConds) != len(ts.AccessCondition) {
-				return errors.New("rebuild range for cached plan failed")
-			}
-			ts.Ranges = res.Ranges
-		} else {
-			ts.Ranges = ranger.FullRange()
-		}
-	} else {
-		var pkCol *expression.Column
-		if ts.Table.PKIsHandle {
-			if pkColInfo := ts.Table.GetPkColInfo(); pkColInfo != nil {
-				pkCol = expression.ColInfo2Col(ts.schema.Columns, pkColInfo)
-			}
-		}
-		if pkCol != nil {
-			ts.Ranges, err = ranger.BuildTableRange(ts.AccessCondition, sctx, pkCol.RetType)
-			if err != nil {
-				return err
-			}
-		} else {
-			ts.Ranges = ranger.FullIntRange(false)
-		}
-	}
-	return
-}
-
-func buildRangeForIndexScan(sctx sessionctx.Context, is *PhysicalIndexScan) (err error) {
-	if len(is.IdxCols) == 0 {
-		is.Ranges = ranger.FullRange()
-		return
-	}
-	res, err := ranger.DetachCondAndBuildRangeForIndex(sctx, is.AccessCondition, is.IdxCols, is.IdxColLens)
-	if err != nil {
-		return err
-	}
-	if len(res.AccessConds) != len(is.AccessCondition) {
-		return errors.New("rebuild range for cached plan failed")
-	}
-	is.Ranges = res.Ranges
-	return
-}
-
-func setFoundInPlanCache(sctx sessionctx.Context, opt bool) error {
-	vars := sctx.GetSessionVars()
-	err := vars.SetSystemVar(variable.TiDBFoundInPlanCache, variable.BoolToOnOff(opt))
-	return err
 }
